Add files via upload
It shows the angle of the Axis in monitor Serial by pressing l to switch on and d to switch off. It makes a sound if the angle value is more than the threshold in axis X. But it possible to make that in any axis. An HC-04 ultrasound sensor is used to detect if there is any object above the device.